:recycle: Merge all sub traits into one
